returns instance of geneformer.Classifier

returns instance of geneformer.Classifier

setup_and_run_Classifier(
  input_data_file,
  output_directory,
  output_prefix,
  split_id_list,
  classifier,
  cell_state_dict,
  filter_data,
  training_args,
  max_ncells,
  freeze_layers,
  num_crossval_splits,
  forward_batch_size,
  nproc,
  base_model_dir,
  val_split_id_list
)

Arguments

input_data_file

character(1) path to a `.datasets` folder loadable by huggingface `datasets` module

output_directory

character(1) folder that will be populated with fine tuning test and train outcome resources. For the heart failure example, 2.2 GB of results are produced, possibly inefficient owing to caching practices.

output_prefix

character(1) tag prepended (or postfixed) to certain generated resources

split_id_list

an R list() with components `attr_key`, `train` and `test`

classifier

character(1) defaults to "cell", see sources of python modules for options (:-()

cell_state_dict

an R list() of cell type tags, with 'name' corresponding to the name used in the tokenized base data (h5ad or loom)

filter_data

a list of e.g., cell type tags, that will be analyzed

training_args

a list with such components as `num_train_epochs` and `learning_rate` -- look fo authoritative doc

max_ncells

integer(1)

freeze_layers

integer(1) seek authoritative doc

num_crossval_splits

integer(1)

forward_batch_size

integer(1)

nproc

integer(1) concerning parallelization

base_model_dir

character(1) path to safetensors representation of model

val_split_id_list

list() with elements `attr_key`, `train`, `eval`

Value

a list with information on accuracy

Note

We pick up the Geneformer package frozen with BiocGF. `token_dictionary.pkl` must be present in working dir