work with a template python script to illustrate mask substitution by a BERT model, and use the pipeline unmasker to produce a list of scored substitutions

work with a template python script to illustrate mask substitution by a BERT model, and use the pipeline unmasker to produce a list of scored substitutions

usebert_pipeline(phrase = "The capital of Spain is [MASK].")

Arguments

phrase

character(1) a string with `[MASK]` ending with a `.`.

Examples

usebert_pipeline()
#> [[1]]
#> [[1]]$score
#> [1] 0.6471179
#> 
#> [[1]]$token
#> [1] 6921
#> 
#> [[1]]$token_str
#> [1] "madrid"
#> 
#> [[1]]$sequence
#> [1] "the capital of spain is madrid."
#> 
#> 
#> [[2]]
#> [[2]]$score
#> [1] 0.1030244
#> 
#> [[2]]$token
#> [1] 7623
#> 
#> [[2]]$token_str
#> [1] "barcelona"
#> 
#> [[2]]$sequence
#> [1] "the capital of spain is barcelona."
#> 
#> 
#> [[3]]
#> [[3]]$score
#> [1] 0.05209291
#> 
#> [[3]]$token
#> [1] 18983
#> 
#> [[3]]$token_str
#> [1] "seville"
#> 
#> [[3]]$sequence
#> [1] "the capital of spain is seville."
#> 
#> 
#> [[4]]
#> [[4]]$score
#> [1] 0.02670261
#> 
#> [[4]]$token
#> [1] 13083
#> 
#> [[4]]$token_str
#> [1] "valencia"
#> 
#> [[4]]$sequence
#> [1] "the capital of spain is valencia."
#> 
#> 
#> [[5]]
#> [[5]]$score
#> [1] 0.02181132
#> 
#> [[5]]$token
#> [1] 27382
#> 
#> [[5]]$token_str
#> [1] "malaga"
#> 
#> [[5]]$sequence
#> [1] "the capital of spain is malaga."
#> 
#>